+++ src/latexgen.h
@@ -316,7 +316,7 @@ class LatexGenerator final : public OutputGenerator, public OutputGenIntf
     void startPlainFile(const DString &name) override { OutputGenerator::startPlainFile(name); }
     void endPlainFile() override { OutputGenerator::endPlainFile(); }
 
-    void startEmbeddedDoc(int) override;
+    void startEmbeddedDoc(size_t) override;
     void endEmbeddedDoc() override;
 
   private:
